start to add adt data type in jit test
