+++ test/compiler/test_jit.cc
@@ -284,55 +284,6 @@ sq 10.0
     engine_free(engine);
 }
 
-TEST(testJIT, testIfFunc)
-{
-    char test_code[] = R"(
-let if_f x = 
-    if x < 10 then x
-    else 0
-if_f 5
-if_f 10
-)";
-    struct engine *engine = engine_llvm_new(false);
-    struct cg_llvm *cg = (struct cg_llvm*)engine->be->cg;
-    JIT *jit = build_jit(engine);
-    struct ast_node *block = parse_code(engine->fe->parser, test_code);
-    analyze(cg->base.sema_context, block);
-    auto node = (struct ast_node *)array_front_ptr(&block->block->nodes);
-    auto node1 = (struct ast_node *)array_get_ptr(&block->block->nodes, 1);
-    auto node2 = (struct ast_node *)array_get_ptr(&block->block->nodes, 2);
-    eval_statement(jit, node);
-    eval_result result1 = eval_exp(jit, node1);
-    eval_result result2 = eval_exp(jit, node2);
-    ASSERT_EQ(5, result1.i_value);
-    ASSERT_EQ(0, result2.i_value);
-    node_free(block);
-    jit_free(jit);
-    engine_free(engine);
-}
-
-TEST(testJIT, testForLoopFunc)
-{
-    char test_code[] = R"(
-let forloop n = 
-    for i in 1..n
-        i
-forloop 4
-  )";
-    struct engine *engine = engine_llvm_new(false);
-    struct cg_llvm *cg = (struct cg_llvm*)engine->be->cg;
-    JIT *jit = build_jit(engine);
-    struct ast_node *block = parse_code(engine->fe->parser, test_code);
-    analyze(cg->base.sema_context, block);
-    auto node = (struct ast_node *)array_front_ptr(&block->block->nodes);
-    auto node1 = (struct ast_node *)array_back_ptr(&block->block->nodes);
-    eval_statement(jit, node);
-    eval_result result1 = eval_exp(jit, node1);
-    ASSERT_EQ(0, result1.i_value);
-    node_free(block);
-    jit_free(jit);
-    engine_free(engine);
-}
 /*TODO: The following operator override is not supported
 TEST(testJIT, testUnaryFunc)
 {
@@ -474,110 +425,3 @@ printf "hello:%d" 1
     engine_free(engine);
 }
 #endif
-
-TEST(testJIT, testProductType)
-{
-    char test_code[] = R"(
-struct Point2D = x:f64, y:f64
-let xy:Point2D = Point2D { 10.0, 20.0 }
-xy.x
-xy.y
-)";
-    struct engine *engine = engine_llvm_new(true);
-    struct cg_llvm *cg = (struct cg_llvm*)engine->be->cg;
-    JIT *jit = build_jit(engine);
-    struct ast_node *block = parse_code(engine->fe->parser, test_code);
-    analyze(cg->base.sema_context, block);
-    auto end = 2;
-    for (int i = 0; i < end; i++) {
-        auto node = (struct ast_node *)array_get_ptr(&block->block->nodes, i);
-        eval_statement(jit, node);
-    }
-    auto node3 = (struct ast_node *)array_get_ptr(&block->block->nodes, 2);
-    ASSERT_EQ(10.0, eval_exp(jit, node3).d_value);
-    auto node4 = (struct ast_node *)array_get_ptr(&block->block->nodes, 3);
-    ASSERT_EQ(20.0, eval_exp(jit, node4).d_value);
-    node_free(block);
-    jit_free(jit);
-    engine_free(engine);
-}
-
-TEST(testJIT, testProductTypeIntType)
-{
-    char test_code[] = R"(
-struct Point2D = x:int, y:int
-let xy:Point2D = Point2D { 10, 20 }
-xy.x
-xy.y
-)";
-    struct engine *engine = engine_llvm_new(true);
-    struct cg_llvm *cg = (struct cg_llvm*)engine->be->cg;
-    JIT *jit = build_jit(engine);
-    struct ast_node *block = parse_code(engine->fe->parser, test_code);
-    analyze(cg->base.sema_context, block);
-    auto end = 2;
-    for (int i = 0; i < end; i++) {
-        auto node = (struct ast_node *)array_get_ptr(&block->block->nodes, i);
-        eval_statement(jit, node);
-    }
-    auto node3 = (struct ast_node *)array_get_ptr(&block->block->nodes, 2);
-    ASSERT_EQ(10, eval_exp(jit, node3).i_value);
-    auto node4 = (struct ast_node *)array_get_ptr(&block->block->nodes, 3);
-    ASSERT_EQ(20, eval_exp(jit, node4).i_value);
-    node_free(block);
-    jit_free(jit);
-    engine_free(engine);
-}
-
-TEST(testJIT, testProductTypeMixedType)
-{
-    char test_code[] = R"(
-struct Point2D = x:f64, y:int
-let xy:Point2D = Point2D { 10.0, 20 }
-xy.x
-xy.y
-)";
-    struct engine *engine = engine_llvm_new(true);
-    struct cg_llvm *cg = (struct cg_llvm*)engine->be->cg;
-    JIT *jit = build_jit(engine);
-    struct ast_node *block = parse_code(engine->fe->parser, test_code);
-    analyze(cg->base.sema_context, block);
-    auto end = 2;
-    for (int i = 0; i < end; i++) {
-        auto node = (struct ast_node *)array_get_ptr(&block->block->nodes, i);
-        eval_statement(jit, node);
-    }
-    auto node3 = (struct ast_node *)array_get_ptr(&block->block->nodes, 2);
-    ASSERT_EQ(10.0, eval_exp(jit, node3).d_value);
-    auto node4 = (struct ast_node *)array_get_ptr(&block->block->nodes, 3);
-    ASSERT_EQ(20, eval_exp(jit, node4).i_value);
-    node_free(block);
-    jit_free(jit);
-    engine_free(engine);
-}
-
-TEST(testJIT, testProductTypeMixedTypeLocalVariable)
-{
-    char test_code[] = R"(
-struct Point2D = x:f64, y:int
-let getx()=
-    let xy:Point2D = Point2D { 10.0, 20 }
-    xy.x
-getx()
-)";
-    struct engine *engine = engine_llvm_new(true);
-    struct cg_llvm *cg = (struct cg_llvm*)engine->be->cg;
-    JIT *jit = build_jit(engine);
-    struct ast_node *block = parse_code(engine->fe->parser, test_code);
-    analyze(cg->base.sema_context, block);
-    auto end = 2;
-    for (int i = 0; i < end; i++) {
-        auto node = (struct ast_node *)array_get_ptr(&block->block->nodes, i);
-        eval_statement(jit, node);
-    }
-    auto node3 = (struct ast_node *)array_get_ptr(&block->block->nodes, 2);
-    ASSERT_EQ(10.0, eval_exp(jit, node3).d_value);
-    node_free(block);
-    jit_free(jit);
-    engine_free(engine);
-}
